Goodbye exec.ts
