+++ packages/core/src/actions/invoke.ts
@@ -3,14 +3,17 @@ import { invoke as invokeActionType } from '../actionTypes';
 import { isActorRef } from '../actors';
 import { createDynamicAction } from '../../actions/dynamicAction';
 import {
+  AnyInterpreter,
+  AnyState,
   BaseDynamicActionObject,
   DynamicInvokeActionObject,
   InvokeActionObject
 } from '..';
-import { actionTypes } from '../actions';
-import { mapContext } from '../utils';
+import { actionTypes, error } from '../actions';
+import { mapContext, warn } from '../utils';
 import { interpret } from '../interpreter';
 import { cloneState } from '../State';
+import { IS_PRODUCTION } from '../environment';
 
 export function invoke<
   TContext extends MachineContext,
@@ -75,6 +78,37 @@ export function invoke<
           [id]: resolvedInvokeAction.params.ref!
         }
       });
+
+      resolvedInvokeAction.execute2 = (actorCtx) => {
+        const interpreter = actorCtx.self as AnyInterpreter;
+        const { id, autoForward, ref } = resolvedInvokeAction.params;
+        if (!ref) {
+          if (!IS_PRODUCTION) {
+            warn(
+              false,
+              `Actor type '${resolvedInvokeAction.params.src.type}' not found in machine '${actorCtx.id}'.`
+            );
+          }
+          return;
+        }
+        ref._parent = interpreter; // TODO: fix
+        actorCtx.defer?.((state: AnyState) => {
+          try {
+            const currentRef = state.children[id];
+            if (!currentRef) {
+              return;
+            }
+            if (autoForward) {
+              interpreter._forwardTo.add(currentRef);
+            }
+
+            currentRef.start?.();
+          } catch (err) {
+            interpreter.send(error(id, err));
+            return;
+          }
+        });
+      };
 
       return [invokedState, resolvedInvokeAction];
     }
